+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## Master
+* Migrate to lsp-protocol.
+
 ## 1.0.0
 * Migrate from `lsp-mode`.